pyproject.toml
[tool.poetry]
name = "drf-les"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"
django-apscheduler = "^0.7.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
